# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 驅動的網紅案件管理系統 API\n提供 Google OAuth 認證、郵件管理、案件管理、AI 分析等功能",
        "title": "Influenter API",
        "termsOfService": "http://influenter.example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://influenter.example.com/support",
            "email": "support@influenter.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/emails": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得使用者的郵件列表，支援分頁、篩選、搜尋",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "郵件"
                ],
                "summary": "取得郵件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth 帳號 ID",
                        "name": "oauth_account_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否已讀",
                        "name": "is_read",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "案件 ID",
                        "name": "case_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "寄件者 email",
                        "name": "from_email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "主旨關鍵字",
                        "name": "subject",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "開始日期 (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "結束日期 (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "頁數",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每頁數量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "received_at",
                        "description": "排序欄位",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "排序方向 (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "郵件列表和分頁資訊",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emails/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得單封郵件的完整內容",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "郵件"
                ],
                "summary": "取得郵件詳情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "郵件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_models.EmailDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新郵件狀態（標記已讀、關聯案件等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "郵件"
                ],
                "summary": "更新郵件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "郵件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新內容",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.UpdateEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_models.EmailDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_designcomb_influenter-backend_internal_models.EmailDetailResponse": {
            "type": "object",
            "properties": {
                "ai_analysis_id": {
                    "type": "string"
                },
                "ai_analyzed": {
                    "type": "boolean"
                },
                "body_html": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "case_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "from_email": {
                    "type": "string"
                },
                "from_name": {
                    "type": "string"
                },
                "has_attachments": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "oauth_account_id": {
                    "type": "string"
                },
                "provider_message_id": {
                    "type": "string"
                },
                "received_at": {
                    "type": "string"
                },
                "snippet": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                },
                "to_email": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.UpdateEmailRequest": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}